SWDEV-459254 - Change return type to support values > 128.


[ROCm/clr commit: c7461436d6d899e6c1625554d59b6c0b859e5707]

+++ rocclr/device/devkernel.hpp
@@ -500,7 +500,7 @@ class Kernel : public amd::HeapObject {
   const uint32_t KernargSegmentByteSize() const { return kernargSegmentByteSize_; }
   void SetKernargSegmentByteSize(uint32_t size) { kernargSegmentByteSize_ = size; }
 
-  const uint8_t KernargSegmentAlignment() const { return kernargSegmentAlignment_; }
+  const uint32_t KernargSegmentAlignment() const { return kernargSegmentAlignment_; }
   void SetKernargSegmentAlignment(uint32_t align) { kernargSegmentAlignment_ = align; }
 
   void SetSymbolName(const std::string& name) { symbolName_ = name; }
